+++ src/record.rs
@@ -7,23 +7,23 @@ pub trait Record: Clone + Index<usize, Output = [u8]> {
     // fn get(&self, index: usize) -> Option<&[u8]>;
     // fn clear(&mut self);
     // fn truncate(&mut self, len: usize);
-    fn push_field(&mut self, cell: &[u8]);
+    // fn push_field(&mut self, cell: &[u8]);
 
     // #[inline(always)]
     // fn is_empty(&self) -> bool {
     //     self.len() == 0
     // }
 
-    #[must_use]
-    fn replace_at<'a>(&'a self, column_index: usize, new_value: &'a [u8]) -> Self
-    where
-        Self: FromIterator<&'a [u8]>,
-    {
-        self.iter()
-            .enumerate()
-            .map(|(i, v)| if i == column_index { new_value } else { v })
-            .collect()
-    }
+    // #[must_use]
+    // fn replace_at<'a>(&'a self, column_index: usize, new_value: &'a [u8]) -> Self
+    // where
+    //     Self: FromIterator<&'a [u8]>,
+    // {
+    //     self.iter()
+    //         .enumerate()
+    //         .map(|(i, v)| if i == column_index { new_value } else { v })
+    //         .collect()
+    // }
 
     // #[must_use]
     // fn prepend<'a>(&'a self, cell_value: &[u8]) -> Self
@@ -37,12 +37,12 @@ pub trait Record: Clone + Index<usize, Output = [u8]> {
     //     new_record
     // }
 
-    #[must_use]
-    fn append(&self, cell_value: &[u8]) -> Self {
-        let mut new_record = self.clone();
-        new_record.push_field(cell_value);
-        new_record
-    }
+    // #[must_use]
+    // fn append(&self, cell_value: &[u8]) -> Self {
+    //     let mut new_record = self.clone();
+    //     new_record.push_field(cell_value);
+    //     new_record
+    // }
 
     #[must_use]
     fn remove<'a>(&'a self, column_index: usize) -> Self
@@ -87,10 +87,10 @@ impl Record for csv::ByteRecord {
     //     self.truncate(len);
     // }
 
-    #[inline(always)]
-    fn push_field(&mut self, cell: &[u8]) {
-        self.push_field(cell);
-    }
+    // #[inline(always)]
+    // fn push_field(&mut self, cell: &[u8]) {
+    //     self.push_field(cell);
+    // }
 }
 
 impl Record for simd_csv::ByteRecord {
@@ -124,8 +124,8 @@ impl Record for simd_csv::ByteRecord {
     //     self.truncate(len);
     // }
 
-    #[inline(always)]
-    fn push_field(&mut self, cell: &[u8]) {
-        self.push_field(cell);
-    }
+    // #[inline(always)]
+    // fn push_field(&mut self, cell: &[u8]) {
+    //     self.push_field(cell);
+    // }
 }