+++ tests/tests.rs
@@ -26,7 +26,6 @@ mod test_enumerate;
 mod test_explode;
 mod test_filter;
 mod test_fixlengths;
-mod test_flatmap;
 mod test_fmt;
 mod test_frequency;
 mod test_fuzzy_join;